+++ ntex-grpc/generate.sh
@@ -2,3 +2,4 @@
 
 ntex-grpc timestamp.proto timestamp.rs --out-dir ./src/google_types --include-dir ../prost-build/third-party/include/google/protobuf/
 ntex-grpc duration.proto duration.rs --out-dir ./src/google_types --include-dir ../prost-build/third-party/include/google/protobuf/
+ntex-grpc wrappers.proto wrappers.rs --out-dir ./src/google_types --include-dir ../prost-build/third-party/include/google/protobuf/