+++ ntex-grpc/src/google_types/mod.rs
@@ -2,6 +2,8 @@ mod duration;
 mod duration_impl;
 mod timestamp;
 mod timestamp_impl;
+mod wrappers;
 
 pub use self::duration::Duration;
 pub use self::timestamp::Timestamp;
+pub use self::wrappers::*;